fix(unit_price): Corrects unit price conversion > R$1,000
When transactions have unit prices greater than R$ 1.000,00 the code
crashes, because it is unable to convert unit prices with "." into a
decimal.

The change ensures that no error occurs when converting the unit price
and amount parse to Decimal now.

Close #5

+++ correpy/domain/entities/transaction.py
@@ -10,7 +10,7 @@ BRAZIL_SOURCE_WITHHELD_TAX_PERCENTAGE = Decimal(0.005)
 @dataclass
 class Transaction:
     transaction_type: TransactionType
-    amount: int
+    amount: Decimal
     unit_price: Decimal
     security: Security
     source_withheld_taxes: Decimal = field(init=False, default=Decimal(0))
